\name{CC.Example.Data1}
\docType{data}
\alias{CC.Example.Data1}
\title{Example Dataset1: Temperature and Flowering Time Series}
\description{
  April mean temperature as well as flowering dates of lilac at the phenological station Niederrimbach, Germany. 
}
\usage{data(CC.Example.Data1)}
\format{A matrix dim=c(60,3) containing 60 observations from 1951 to 2010.}
\source{Oesterle et al. (2006): Qualitaetspruefung, Ergaenzung und Homogenisierung der taeglichen Datenreihen in Deutschland, 1951-2003: ein neuer Datensatz; and www.dwd.de}
\references{
Oesterle et al. (2006): Qualitaetspruefung, Ergaenzung und Homogenisierung der taeglichen Datenreihen in Deutschland, 1951-2003: ein neuer Datensatz; and www.dwd.de
}
\keyword{datasets}