\name{CC.Example.Data2}
\docType{data}
\alias{CC.Example.Data2}
\title{Example Dataset2: Soil and crop data from Kenya}
\description{
  Information of soil organic carbon contents, sorghum plantation and intercropping on 218 plots in western Kenya. 
}
\usage{data(CC.Example.Data2)}
\format{A matrix dim=c(218,4) containing observations for 218 plots.}
\source{http://samples.ccafs.cgiar.org}
\references{http://samples.ccafs.cgiar.org}
\keyword{datasets}
